set cookie for sidebar preference
